+++ acm_web_site/apps/business/activities/admin.py
@@ -16,6 +16,7 @@ from .models import (
 
 # Register your models here.
 
+
 class ActivityAdmin(admin.ModelAdmin):
     form = ActivityAdminForm
