--- acm_web_site/apps/business/web_site/admin.py
@@ -1,115 +1,0 @@
-from django.contrib import admin
-
-from django.core.files.storage import FileSystemStorage
-
-from .models import *
-from .forms import (
-	GalleryAdminForm,
-	AwardAdminForm,
-	ActivityAdminForm,
-	MemberAdminForm,
-	ProjectAdminForm,
-	TeamAdminForm,
-	TutorialAdminForm
-)
-
-
-class GalleryAdmin(admin.ModelAdmin):
-	form = GalleryAdminForm
-
-	def save_model(self, request, obj, form, change):
-		file_form = request.FILES['picture']
-		fs = FileSystemStorage()
-		filename = fs.save(file_form.name, file_form)
-		uploaded_file_url = fs.url(filename)
-		obj.picture = uploaded_file_url
-		super().save_model(request, obj, form, change)
-
-
-class AwardAdmin(admin.ModelAdmin):
-	form = AwardAdminForm
-
-	def save_model(self, request, obj, form, change):
-		file_form = request.FILES['picture']
-		fs = FileSystemStorage()
-		filename = fs.save(file_form.name, file_form)
-		uploaded_file_url = fs.url(filename)
-		obj.picture = uploaded_file_url
-		super().save_model(request, obj, form, change)
-
-
-class ActivityAdmin(admin.ModelAdmin):
-	form = ActivityAdminForm
-
-	def save_model(self, request, obj, form, change):
-		file_form = request.FILES['poster']
-		fs = FileSystemStorage()
-		filename = fs.save(file_form.name, file_form)
-		uploaded_file_url = fs.url(filename)
-		obj.poster = uploaded_file_url
-		super().save_model(request, obj, form, change)
-
-
-class MemberAdmin(admin.ModelAdmin):
-	form = MemberAdminForm
-
-	def save_model(self, request, obj, form, change):
-		file_form = request.FILES['picture']
-		fs = FileSystemStorage()
-		filename = fs.save(file_form.name, file_form)
-		uploaded_file_url = fs.url(filename)
-		obj.picture = uploaded_file_url
-		super().save_model(request, obj, form, change)
-
-
-class ProjectAdmin(admin.ModelAdmin):
-	form = ProjectAdminForm
-
-	def save_model(self, request, obj, form, change):
-		file_form = request.FILES['poster']
-		fs = FileSystemStorage()
-		filename = fs.save(file_form.name, file_form)
-		uploaded_file_url = fs.url(filename)
-		obj.poster = uploaded_file_url
-		super().save_model(request, obj, form, change)
-
-
-class TeamAdmin(admin.ModelAdmin):
-	form = TeamAdminForm
-
-	def save_model(self, request, obj, form, change):
-		file_form = request.FILES['picture']
-		fs = FileSystemStorage()
-		filename = fs.save(file_form.name, file_form)
-		uploaded_file_url = fs.url(filename)
-		obj.picture = uploaded_file_url
-		super().save_model(request, obj, form, change)
-
-
-class TutorialAdmin(admin.ModelAdmin):
-	form = TutorialAdminForm
-
-	def save_model(self, request, obj, form, change):
-		file_form = request.FILES['poster']
-		fs = FileSystemStorage()
-		filename = fs.save(file_form.name, file_form)
-		uploaded_file_url = fs.url(filename)
-		obj.poster = uploaded_file_url
-		super().save_model(request, obj, form, change)
-
-
-admin.site.register(Contest)
-admin.site.register(Member, MemberAdmin)
-admin.site.register(Team, TeamAdmin)
-admin.site.register(Award, AwardAdmin)
-admin.site.register(Gallery, GalleryAdmin)
-admin.site.register(Tutorial, TutorialAdmin)
-admin.site.register(Activity, ActivityAdmin)
-admin.site.register(Project, ProjectAdmin)
-# admin.site.register(TeamContest)
-# admin.site.register(TutorialFile)
-# admin.site.register(ActivityFile)
-# admin.site.register(ActivityMember)
-# admin.site.register(ProjectFile)
-# admin.site.register(ProjectMember)
-# admin.site.register(TeamMember)
