--- acm_web_site/apps/business/web_site/api/award_service.py
@@ -1,5 +1,0 @@
-from ..models import Award
-
-
-def get_awards():
-	return Award.objects.all()
