--- acm_web_site/apps/business/web_site/api/project_service.py
@@ -1,16 +1,0 @@
-from ..models import Project
-
-
-def get_projects():
-    return Project.objects.all()
-
-
-def get_project(id):
-    """
-    :param id: int
-    :return: Project
-    """
-    try:
-        return Project.objects.get(id=id)
-    except Project.DoesNotExist:
-        return None
