--- acm_web_site/apps/business/web_site/api/team_service.py
@@ -1,8 +1,0 @@
-from ..models import Team
-
-
-def get_teams():
-    """
-    :return: QuerySet<Team>
-    """
-    return Team.objects.all()
