--- acm_web_site/apps/business/web_site/api/tutorial_service.py
@@ -1,16 +1,0 @@
-from ..models import Tutorial
-
-
-def get_tutorials():
-    return Tutorial.objects.all()
-
-
-def get_tutorial(id):
-    """
-    :param id: int
-    :return: Project
-    """
-    try:
-        return Tutorial.objects.get(id=id)
-    except Tutorial.DoesNotExist:
-        return None
