--- acm_web_site/apps/business/web_site/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class WebsiteConfig(AppConfig):
-    name = 'web_site'
