--- acm_web_site/apps/business/web_site/serializer/team_serializer.py
@@ -1,23 +1,0 @@
-# -- coding: utf-8
-from __future__ import unicode_literals
-import base64
-
-from rest_framework import serializers
-
-from django.conf import settings
-from ..models import Team
-
-
-class TeamSerializer(serializers.ModelSerializer):
-    picture = serializers.SerializerMethodField(required=False)
-
-    class Meta:
-        model = Team
-        fields = '__all__'
-
-    def get_picture(self, obj):
-        prefix = '/'.join(settings.MEDIA_ROOT.split('/')[:-1])
-        complete_path = prefix + obj.picture
-        with open(complete_path, "rb") as image_file:
-            str = base64.b64encode(image_file.read())
-        return str
