+++ acm_web_site/settings/common.py
@@ -35,7 +35,6 @@ THIRD_PARTY_APPS = [
 
 LOCAL_APPS = [
     'apps.utils',
-    'apps.business.web_site',
     'apps.business.activities',
 ]
