+++ tox.ini
@@ -1,4 +1,4 @@
 [flake8]
 ignore = E402, F401, F403, F405, W191
 max-line-length = 79
-exclude = */migrations/*.py, .environment, node_modules, ENV
+exclude = */migrations/*.py, .environment, node_modules, ENV, venv
